src/main/java/com/girigiri/kwrental/equipment/service/EquipmentService.java
package com.girigiri.kwrental.equipment.service;

import com.girigiri.kwrental.equipment.Equipment;
import com.girigiri.kwrental.equipment.EquipmentRepository;
import com.girigiri.kwrental.equipment.dto.EquipmentResponse;
import com.girigiri.kwrental.equipment.exception.EquipmentNotFoundException;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

@Service
public class EquipmentService {

    private final EquipmentRepository equipmentRepository;

    public EquipmentService(final EquipmentRepository equipmentRepository) {
        this.equipmentRepository = equipmentRepository;
    }

    @Transactional(readOnly = true)
    public EquipmentResponse findById(final Long id) {
        final Equipment equipment = equipmentRepository.findById(id)
                .orElseThrow(EquipmentNotFoundException::new);
        return EquipmentResponse.from(equipment);
    }
}
